requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart

numpy
faiss-cpu

pypdf
python-docx
pdfplumber
pymupdf
pillow